Certificate For Contribution
I'm not sure if this is the right place to place this certificate and it's public view of my contributing to the web as a self learning as I go and/or can music metadata bookmarking (with indie artists at heart and mind) beginner developer. I'm not the smartest of all so it's taking me a bit to grasp the understanding of what I'm trying to learn but I'm not a quitter and I know that God is gonna help me do this because ultimately I was created to fulfill His purpose for I trust in Him and He said that His plans are good to prosper me and not to hurt me. Praise, glory, honor, and worship to Him alone if I accomplish anything because I couldn't have done anything alone. Amen